Take authors on command-line.
When doing things programatically, it's annoying to have to do input on a
prompt; even when doing things interactively, it still is. :)

This allows us to enter some number of authors after the license name;
if it's zero, we'll still prompt like before.

Since the author is part of every single license, pull out the prompting
one level to licgen itself, rather than erb.